[FIX] merge inherits
