+++ l10n_cl_financial_indicators/models/l10n_cl_financial_indicators.py
@@ -2,6 +2,7 @@ import logging
 
 import requests
 from odoo import api, fields, models
+from odoo.addons import decimal_precision as dp
 
 _logger = logging.getLogger(__name__)
 
@@ -15,6 +16,36 @@ indicators = {
 class ResCurrency(models.Model):
     _name = "res.currency"
     _inherit = "res.currency"
+
+    rate = fields.Float(
+        compute='_compute_current_rate', string='Current Rate',
+        digits=dp.get_precision('Currency'),
+        help='The rate of the currency to the currency of rate 1.')
+
+    rounding = fields.Float(
+        string='Rounding Factor', default=0.01,
+        digits=dp.get_precision('Currency'))
+
+    @api.multi
+    def _compute_current_rate(self):
+        date = self._context.get('date') or fields.Datetime.now()
+        company_id = self._context.get(
+            'company_id') or self.env['res.users']._get_company().id
+        # the subquery selects the last rate before 'date' for
+        # the given currency/company
+        query = """SELECT c.id, (
+        SELECT r.rate
+        FROM res_currency_rate r
+        WHERE r.currency_id = c.id AND r.name <= %s
+        AND (r.company_id IS NULL OR r.company_id = %s)
+        ORDER BY r.company_id, r.name DESC
+        LIMIT 1) AS rate
+            FROM res_currency c
+            WHERE c.id IN %s"""
+        self._cr.execute(query, (date, company_id, tuple(self.ids)))
+        currency_rates = dict(self._cr.fetchall())
+        for currency in self:
+            currency.rate = currency_rates.get(currency.id) or 1.0
 
     @api.model
     def update_currency(self, indic, currency_id=False):
